+++ pysui/sui/sui_builders/exec_builders.py
@@ -41,6 +41,7 @@ class ExecuteTransaction(_NativeTransactionBuilder):
         "showEvents": True,
         "showInput": True,
         "showObjectChanges": True,
+        "showRawInput": True,
     }
 
     @sui_builder()